+++ R/measures.R
@@ -241,6 +241,10 @@ rsq = makeMeasure(id = "rsq", minimize = FALSE, best = 1, worst = -Inf,
 measureRSQ = function(truth, response) {
   rss = measureSSE(truth, response)
   ess = sum((truth - mean(truth))^2L)
+  if (ess == 0){
+    warning("Measure is undefined if all truth values are equal.")
+    return(NA_real_)
+  }
   1 - rss / ess
 }
 
@@ -262,6 +266,10 @@ expvar = makeMeasure(id = "expvar", minimize = FALSE, best = 1, worst = 0,
 measureEXPVAR = function(truth, response) {
   regss = sum((response - mean(truth))^2L)
   ess = sum((truth - mean(truth))^2L)
+  if (ess == 0){
+    warning("Measure is undefined if all truth values are equal.")
+    return(NA_real_)
+  }
   regss / ess
 }
 
@@ -275,6 +283,10 @@ arsq = makeMeasure(id = "adjrsq", minimize = FALSE, best = 1, worst = 0,
   fun = function(task, model, pred, feats, extra.args) {
     n = length(pred$data$truth)
     p = length(model$features)
+    if (n == p + 1){
+      warning("Adjusted R-squared is undefined if the number observations is equal to the number of independent variables plus one.")
+      return(NA_real_)
+    }
     1 - (1 - measureRSQ(pred$data$truth, pred$data$response)) * (p / (n - p - 1L))
   }
 )
@@ -297,8 +309,8 @@ rrse = makeMeasure(id = "rrse", minimize = TRUE, best = 0, worst = Inf,
 measureRRSE = function(truth, response){
   tss = sum((truth-mean(truth))^2L)
   if (tss == 0){
-    warning("RAE is undefined if all truth values are equal.")
-    return(NA)
+    warning("Measure is undefined if all truth values are equal.")
+    return(NA_real_)
   }
   sqrt(measureSSE(truth, response)/tss)
 }
@@ -321,8 +333,8 @@ rae = makeMeasure(id = "rae", minimize = TRUE, best = 0, worst = Inf,
 measureRAE = function(truth, response){
   meanad = sum(abs(truth-mean(truth)))
   if (meanad == 0){
-    warning("RAE is undefined if all truth values are equal.")
-    return(NA)
+    warning("Measure is undefined if all truth values are equal.")
+    return(NA_real_)
   }
   return(measureSAE(truth, response)/meanad)
 }
@@ -344,8 +356,8 @@ mape = makeMeasure(id = "mape", minimize = TRUE, best = 0, worst = Inf,
 #' @format none
 measureMAPE = function(truth, response){
   if (any(truth == 0)){
-    warning("MAPE is undefined if any truth value is equal to 0.")
-    return(NA)
+    warning("Measure is undefined if any truth value is equal to 0.")
+    return(NA_real_)
   }
   return(mean(abs((truth-response)/truth)))
 }
@@ -504,6 +516,10 @@ multiclass.aunu = makeMeasure(id = "multiclass.aunu", minimize = FALSE, best = 1
 #' @rdname measures
 #' @format none
 measureAUNU = function(probabilities, truth) {
+  if(length(unique(truth)) != nlevels(truth)){
+    warning("Measure is undefined if there isn't at least one sample per class.")
+	return(NA_real_)
+  }
   mean(vnapply(1:nlevels(truth), function(i) colAUC(probabilities[, i], truth == levels(truth)[i])))
 }
 
@@ -523,6 +539,10 @@ multiclass.aunp = makeMeasure(id = "multiclass.aunp", minimize = FALSE, best = 1
 #' @rdname measures
 #' @format none
 measureAUNP = function(probabilities, truth) {
+  if(length(unique(truth)) != nlevels(truth)){
+    warning("Measure is undefined if there isn't at least one sample per class.")
+	return(NA_real_)
+  }
   sum(vnapply(1:nlevels(truth), function(i) mean(truth == levels(truth)[i]) * colAUC(probabilities[,i], truth == levels(truth)[i])))
 }
 
@@ -1016,19 +1036,36 @@ measureFNR = function(truth, response, negative, positive) {
 ppv = makeMeasure(id = "ppv", minimize = FALSE, best = 1, worst = 0,
   properties = c("classif", "req.pred", "req.truth"),
   name = "Positive predictive value",
-  note = "Defined as: tp / (tp + number of fp). Also called precision.",
+  note = "Defined as: tp / (tp + number of fp). Also called precision. If the denominator is 0, PPV is set to be either 1 or 0 depending on whether the highest probability prediction is positive (1) or negative (0).",
   fun = function(task, model, pred, feats, extra.args) {
-    measurePPV(pred$data$truth, pred$data$response, pred$task.desc$positive)
+    if(pred$predict.type == "prob") {
+      prob = getPredictionProbabilities(pred)
+    } else {
+      prob = NULL
+      }
+    measurePPV(pred$data$truth, pred$data$response, pred$task.desc$positive, prob)
   }
 )
 
 #' @export measurePPV
 #' @rdname measures
 #' @format none
-measurePPV = function(truth, response, positive) {
-  measureTP(truth, response, positive) / sum(response == positive)
+
+measurePPV = function(truth, response, positive, probabilities = NULL) {
+  denominator = sum(response == positive)
+  ifelse(denominator == 0, measureEdgeCase(truth, positive, probabilities), measureTP(truth, response, positive) / denominator)
+} 
+measureEdgeCase = function(truth, positive, prob) {
+  if (!is.null(prob)) {
+    rs = sort(prob, index.return = TRUE)
+    erst = ifelse(truth[getLast(rs$ix)] == positive, 1, 0)
+  } else {
+    erst = NA
+  }
+  erst
 }
 
+  
 #' @export npv
 #' @rdname measures
 #' @format none